Add base tag in iframe.
This is intended to solve #2, and it does stop the bizarre splitting
behavior.  But the links don't work. Console gives this message:

> Unsafe JavaScript attempt to initiate navigation for frame with URL
'http://localhost:8000/' from frame with URL 'about:srcdoc'. The frame
attempting navigation of the top-level window is sandboxed, but the
'allow-top-navigation-by-user-activation' flag is not set and navigation
is not triggered by user activation.

Adding allow-top-navigation-by-user-activation removes this message
but doesn't make links work.